refactor: 예약 버튼 (예약, 취소) 컴포넌트
